+++ gitmesh-controller-server/src/main/scala/com/github/takezoe/gitmesh/controller/util/syntax.scala
@@ -1,4 +1,4 @@
-package com.github.takezoe.gitmesh.controller
+package com.github.takezoe.gitmesh.controller.util
 
 object syntax {
